// src/models/Product.ts
export type Product = {
  CODIGO: string;
  DESCRIPCION: string;
  DESC_ADICIONAL?: string | null;
  RUBRO?: string | null;
  SUB_RUBRO?: string | null;
  SUCURSAL?: string | null;
  SALDO?: number | null;
  DEPOSITO?: string | null;
  PENDIENTE?: number | null;
  LISTA_1?: number | null;
  LISTA_2?: number | null;
  LISTA_3?: number | null;
  LISTA_4?: number | null;
  LISTA_6?: number | null;
};